clear documentation
