+++ AppointmentsManager/ClientApp/src/Footer/Footer.js
