+++ AppointmentsManager/ClientApp/src/Footer/footer.css
