+++ AppointmentsManager/ClientApp/src/Login Forms/Login.js
@@ -1,7 +1,8 @@
-﻿import React, { useState } from 'react';
+﻿import './login_rejister.css'; // Import the CSS file
+import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
-import './login_rejister.css'; // Import the CSS file
-import backgroundImage from './assest/Background1.png'; // Import your background image
+import backgroundImage from '../assest/Background1.png'; // Import your background image
+
 function Login() {
     const [loginName, setLoginName] = useState('');
     const [password, setPassword] = useState('');
@@ -26,7 +27,7 @@ function Login() {
         }).then((result) => {
             result.json().then((resp) => {
                 console.warn(resp);
-                navigate("/userDrawingHall");
+                navigate("/user");
             })
         }).catch((error) => {
             // Handle any network errors or other exceptions here
@@ -35,6 +36,11 @@ function Login() {
     }
 
     return (
+
+
+    
+
+
         <div className="homepage" style={{ backgroundImage: `url(${backgroundImage})` }}>
         <div className="bg-gradient-primary">
 
@@ -104,8 +110,8 @@ function Login() {
 
             </div>
 
-          
-            </div> 
+            </div>
+            
 
         </div>
     )
