+++ AppointmentsManager/ClientApp/src/Login Forms/Register.js
@@ -2,7 +2,7 @@
 import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import './login_rejister.css'; // Import the CSS file
-import backgroundImage from './assest/Background1.png'; // Import your background image
+import backgroundImage from '../assest/Background1.png'; // Import your background image
 function Register() {
     const [id, setId] = useState('0');
     const [loginName, setLoginName] = useState('');
