+++ AppointmentsManager/ClientApp/src/Login Forms/staffLogin.js
@@ -1,7 +1,7 @@
 ﻿import React, { useState } from 'react';
 import { useNavigate } from 'react-router-dom';
 import './login_rejister.css'; // Import the CSS file
-import backgroundImage from './assest/Background1.png'; // Import your background image
+import backgroundImage from '../assest/Background1.png'; // Import your background image
 function Login() {
     const [loginName, setLoginName] = useState('');
     const [password, setPassword] = useState('');
@@ -10,7 +10,7 @@ function Login() {
     function GetLoginDetails() {
         // Define admin credentials
         const adminCredentials = [
-            { loginName: 'staff', password: 'staff123', route: '/staffDrawingHall' },
+            { loginName: 'staff', password: 'staff123', route: '/staff' },
       
         ];
 
@@ -70,7 +70,7 @@ function Login() {
                 </div>
             </div>
         </div>
-           </div>
+        </div>
     )
 }
