+++ AppointmentsManager/ClientApp/src/NavBar/NavBar.css
