+++ AppointmentsManager/ClientApp/src/NavBar/NavBar.js
@@ -1,7 +1,7 @@
 ﻿import React from "react";
 import "./NavBar.css"; // Import the CSS file for styling (create this file)
 import { createRoot } from 'react-dom/client';
-import HomeLogo from "./assest/HomeLogo.png"; // Import HomeIcon image
+import HomeLogo from "../assest/HomeLogo.png"; // Import HomeIcon image
 
 
 const rootElement = document.getElementById('root');
