+++ AppointmentsManager/ClientApp/src/components/AllBookingUserComHall/Delete.js
