+++ AppointmentsManager/ClientApp/src/components/AllBookingUserDrawingHall/Home.js
@@ -7,7 +7,7 @@ import New from "./New"
 import Appointment from "./Appointment"
 
 
-import "../../NavBar.css"; // Import the CSS file for styling (create this file)
+import "../../NavBar/NavBar.css"; // Import the CSS file for styling (create this file)
 import "../../custom.css";
 import { getDefault, openModal, filter, getAppointmentsDrawingHall, notifyUser } from "./Lib"
