+++ AppointmentsManager/ClientApp/src/components/ComAdmin/AR.css
