+++ AppointmentsManager/ClientApp/src/components/ComAdmin/Delete.js
