+++ AppointmentsManager/ClientApp/src/components/ComAdmin/DeleteP.js
