+++ AppointmentsManager/ClientApp/src/components/ComAdmin/Edit.js
