+++ AppointmentsManager/ClientApp/src/components/ComAdmin/Lib.js
