+++ AppointmentsManager/ClientApp/src/components/SelectHall/SelectHallPage.js
@@ -1,0 +1,1 @@
+﻿
