+++ AppointmentsManager/ClientApp/src/components/StaffComHall/Delete.js
