+++ AppointmentsManager/ClientApp/src/components/UserComHall/Delete.js
