+++ AppointmentsManager/ClientApp/src/components/UserComHall/Lib.js
