+++ AppointmentsManager/ClientApp/src/components/UserComHall/New.js
